+++ ESMF/GOCART2G_GridComp/SU2G_GridComp/SU2G_GridCompMod.F90
@@ -19,6 +19,8 @@ module SU2G_GridCompMod
    use GOCART2G_Process       ! GOCART2G process library
    use GA_EnvironmentMod
    use MAPL_StringTemplate, only: StrTemplate
+   use MAPL_PackedTimeMod, only: MAPL_PackedDateCreate => PackedDateCreate, &
+                                  MAPL_PackedTimeCreate => PackedTimeCreate
    !$ use omp_lib
 
    implicit none
@@ -847,8 +849,8 @@ contains
 !   ---------------------------------
     call ESMF_ClockGet (clock, currTime=time, __RC__)
     call ESMF_TimeGet (time ,YY=iyr, MM=imm, DD=idd, H=ihr, M=imn, S=isc, __RC__)
-    call MAPL_PackTime (nymd, iyr, imm , idd)
-    call MAPL_PackTime (nhms, ihr, imn, isc)
+    nymd = MAPL_PackedDateCreate(iyr, imm, idd)
+    nhms = MAPL_PackedTimeCreate(ihr, imn, isc)
 
 !   Reset tracer to zero at 0Z on specific day of week
 !   --------------------------------------------------
@@ -1134,8 +1136,8 @@ contains
 !   ---------------------------------
     call ESMF_ClockGet (clock, currTime=time, __RC__)
     call ESMF_TimeGet (time ,YY=iyr, MM=imm, DD=idd, H=ihr, M=imn, S=isc, __RC__)
-    call MAPL_PackTime (nymd, iyr, imm , idd)
-    call MAPL_PackTime (nhms, ihr, imn, isc)
+    nymd = MAPL_PackedDateCreate(iyr, imm, idd)
+    nhms = MAPL_PackedTimeCreate(ihr, imn, isc)
 
 !   Get my private internal state
 !   ------------------------------